# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a ERP/POS API for G4H.",
        "title": "G4H ERP/POS API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "G4H ERP/POS API",
            "url": "https://github.com/aslon1213/g4h_pos_erp",
            "email": "hamidovaslon1@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/admin/suppliers/{branch_id}/{supplier_id}/transactions": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new transaction for a supplier and update financial records",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "suppliers",
                    "transactions"
                ],
                "summary": "Create a new transaction for a supplier",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Branch ID",
                        "name": "branch_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Supplier ID",
                        "name": "supplier_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Transaction data",
                        "name": "transaction",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.TransactionBase"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Output-models_Transaction"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorOutput"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorOutput"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "models.ErrorOutput": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {}
                },
                "error": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Error"
                    }
                }
            }
        },
...
        "models.InitiatorType": {
            "type": "string",
            "enum": [
                "salary",
                "rent",
                "utilities",
                "other",
                "sale",
                "supplier",
                "bnpl"
            ],
            "x-enum-comments": {
                "InitiatorTypeBNPL": "buy now pay later BNPL transactions"
            },
            "x-enum-varnames": [
                "InitiatorTypeSalary",
                "InitiatorTypeRent",
                "InitiatorTypeUtilities",
                "InitiatorTypeOther",
                "InitiatorTypeSales",
                "InitiatorTypeSupplier",
                "InitiatorTypeBNPL"
            ]
        },
...
        "models.Output-models_Transaction": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.Transaction"
                },
                "error": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Error"
                    }
                }
            }
        },
        "models.PaymentMethod": {
            "type": "string",
            "enum": [
                "cash",
                "bank",
                "terminal",
                "online_payment",
                "cheque",
                "online_transfer",
                "undefined"
            ],
            "x-enum-varnames": [
                "PaymentMethodCash",
                "PaymentMethodBank",
                "PaymentMethodTerminal",
                "OnlineMobileAppPayment",
                "Cheque",
                "OnlineTransfer",
                "PaymentMethodUndefined"
            ]
        },
...
        "models.Transaction": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "branch_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "payment_method": {
                    "$ref": "#/definitions/models.PaymentMethod"
                },
                "type": {
                    "$ref": "#/definitions/models.InitiatorType"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.TransactionBase": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "payment_method": {
                    "$ref": "#/definitions/models.PaymentMethod"
                },
                "type": {
                    "$ref": "#/definitions/models.TransactionType"
                }
            }
        },
...
        "models.TransactionType": {
            "type": "string",
            "enum": [
                "credit",
                "debit"
            ],
            "x-enum-comments": {
                "TransactionTypeCredit": "credit means - income - when money is gained or received into an account",
                "TransactionTypeDebit": "debit means - outcome - when money is lost, spent, or withdrawn from an account"
            },
            "x-enum-varnames": [
                "TransactionTypeCredit",
                "TransactionTypeDebit"
            ]
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}